# docs/swagger.yaml
basePath: /
definitions:
...
  api.CreateVirtualServerRequestDto:
    properties:
      additionalSigningAlgorithms:
        items:
          type: string
        type: array
      admin:
        $ref: '#/definitions/api.CreateVirtualServerRequestDtoAdminDto'
      displayName:
        maxLength: 255
        minLength: 1
        type: string
      enableRegistration:
        type: boolean
      name:
        maxLength: 255
        minLength: 1
        type: string
      primarySigningAlgorithm:
        enum:
        - RS256
        - EdDSA
        type: string
      projects:
        items:
          $ref: '#/definitions/api.CreateVirtualServerRequestDtoProjectDto'
        type: array
      require2fa:
        type: boolean
      serviceUsers:
        items:
          $ref: '#/definitions/api.CreateVirtualServerRequestDtoServiceUserDto'
        type: array
    required:
    - displayName
    - name
    type: object
  api.CreateVirtualServerRequestDtoAdminDto:
    properties:
      displayName:
        maxLength: 255
        minLength: 1
        type: string
      passwordHash:
        type: string
      primaryEmail:
        type: string
      roles:
        items:
          type: string
        type: array
      username:
        maxLength: 255
        minLength: 1
        type: string
    required:
    - displayName
    - passwordHash
    - primaryEmail
    - username
    type: object
  api.CreateVirtualServerRequestDtoProjectDto:
    properties:
      applications:
        items:
          $ref: '#/definitions/api.CreateVirtualServerRequestDtoProjectDtoApplicationDto'
        type: array
      description:
        type: string
      name:
        maxLength: 255
        minLength: 1
        type: string
      resourceServers:
        items:
          $ref: '#/definitions/api.CreateVirtualServerRequestDtoProjectDtoResourceServerDto'
        type: array
      roles:
        items:
          $ref: '#/definitions/api.CreateVirtualServerRequestDtoProjectDtoRoleDto'
        type: array
      slug:
        maxLength: 255
        minLength: 1
        type: string
    required:
    - name
    - slug
    type: object
  api.CreateVirtualServerRequestDtoProjectDtoApplicationDto:
    properties:
      displayName:
        maxLength: 255
        minLength: 1
        type: string
      hashedSecret:
        type: string
      name:
        maxLength: 255
        minLength: 1
        type: string
      postLogoutUris:
        items:
          type: string
        type: array
      publicKeys:
        items:
          $ref: '#/definitions/api.CreateVirtualServerRequestDtoProjectDtoApplicationKeyDto'
        type: array
      redirectUris:
        items:
          type: string
        type: array
      tokenEndpointAuthMethod:
        enum:
        - client_secret
        - private_key_jwt
        type: string
      trustedExchangers:
        items:
          type: string
        type: array
      type:
        enum:
        - public
        - confidential
        type: string
      userinfoInAccessToken:
        type: boolean
    required:
    - displayName
    - name
    - redirectUris
    - type
    type: object
  api.CreateVirtualServerRequestDtoProjectDtoApplicationKeyDto:
    properties:
      kid:
        type: string
      pem:
        type: string
    required:
    - kid
    - pem
    type: object
  api.CreateVirtualServerRequestDtoProjectDtoResourceServerDto:
    properties:
      description:
        type: string
      name:
        maxLength: 255
        minLength: 1
        type: string
      slug:
        maxLength: 255
        minLength: 1
        type: string
    required:
    - name
    - slug
    type: object
  api.CreateVirtualServerRequestDtoProjectDtoRoleDto:
    properties:
      description:
        type: string
      name:
        maxLength: 255
        minLength: 1
        type: string
    required:
    - name
    type: object
  api.CreateVirtualServerRequestDtoServiceUserDto:
    properties:
      publicKey:
        properties:
          kid:
            type: string
          pem:
            type: string
        required:
        - kid
        - pem
        type: object
      roles:
        items:
          type: string
        type: array
      username:
        maxLength: 255
        minLength: 1
        type: string
    required:
    - publicKey
    - username
    type: object
...
  api.GetVirtualServerListResponseDto:
    properties:
      displayName:
        type: string
      name:
        type: string
      registrationEnabled:
        type: boolean
    type: object
  api.GetVirtualServerResponseDto:
    properties:
      additionalSigningAlgorithms:
        items:
          type: string
        type: array
      createdAt:
        type: string
      displayName:
        type: string
      id:
        type: string
      name:
        type: string
      primarySigningAlgorithm:
        type: string
      registrationEnabled:
        type: boolean
      require2fa:
        type: boolean
      requireEmailVerification:
        type: boolean
      updatedAt:
        type: string
    type: object
...
  api.PatchVirtualServerRequestDto:
    properties:
      additionalSigningAlgorithms:
        items:
          type: string
        type: array
      displayName:
        type: string
      enableRegistration:
        type: boolean
      primarySigningAlgorithm:
        enum:
        - RS256
        - EdDSA
        type: string
      require2fa:
        type: boolean
      requireEmailVerification:
        type: boolean
    type: object
...
info:
  contact: {}
  description: Open source OIDC/IDP server.
  title: Keyline API
paths:
  /api/virtual-servers:
    post:
      consumes:
      - application/json
      parameters:
      - description: Virtual server
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.CreateVirtualServerRequestDto'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      summary: Create virtual server
      tags:
      - Admin
  /api/virtual-servers/{virtualServerName}:
    get:
      parameters:
      - default: keyline
        description: Virtual server name
        in: path
        name: virtualServerName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.GetVirtualServerResponseDto'
        "404":
          description: Not Found
          schema:
            type: string
      summary: Get virtual server
      tags:
      - Admin
    patch:
      consumes:
      - application/json
      parameters:
      - default: keyline
        description: Virtual server name
        in: path
        name: virtualServerName
        required: true
        type: string
      - description: Patch document
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.PatchVirtualServerRequestDto'
      produces:
      - text/plain
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      summary: Patch virtual server
      tags:
      - Admin
...
  /api/virtual-servers/{virtualServerName}/public-info:
    get:
      parameters:
      - default: keyline
        description: Virtual server name
        in: path
        name: virtualServerName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.GetVirtualServerListResponseDto'
        "404":
          description: Not Found
          schema:
            type: string
      summary: Get virtual server public info
      tags:
      - Admin
...
securityDefinitions:
  BasicAuth:
    type: basic
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"